Add ThirtyPercentDiscountScenario

+++ src/org/dev/Product.java
@@ -4,9 +4,9 @@ public class Product {
 
 	private int id ;
 	private String name ;
-	private int price;
+	private double price;
 	
-	public Product(int id, String name, int price) {
+	public Product(int id, String name, double price) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -29,11 +29,11 @@ public class Product {
 		this.name = name;
 	}
 
-	public int getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(int price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
